+++ cuda_lin_alg/src/gpu_matrix.cuh
@@ -4,7 +4,8 @@
 
 namespace cuda_lin_alg {
 
-lin_alg::Matrix
-        cuda_tiled_multiply(const lin_alg::Matrix& a, const lin_alg::Matrix& b, size_t tile_size);
+lin_alg::Matrix cuda_tiled_multiply(const lin_alg::Matrix& a,
+        const lin_alg::Matrix& b,
+        unsigned int tile_size);
 
 }
